requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
